refactor: moved internal functions to pkg and updated refs
