+++ internal/cache/sqlite/sqlite.go
@@ -3,8 +3,8 @@ package sqlite
 import (
 	"fmt"
 
-	magellan "github.com/OpenCHAMI/magellan/internal"
 	"github.com/OpenCHAMI/magellan/internal/util"
+	magellan "github.com/OpenCHAMI/magellan/pkg"
 
 	"github.com/jmoiron/sqlx"
 )
