+++ pkg/scan.go
